ai_agent/tests/test_web_browser_integration.py
```python
import unittest
from ai_agent import web_browser_integration

class TestWebBrowserIntegration(unittest.TestCase):

    def setUp(self):
        self.webSessionData = {
            "currentUrl": "https://www.example.com",
            "formInputs": {"username": "testuser", "password": "testpass"},
            "cookies": {"session": "123456"}
        }

    def test_navigateWebPage(self):
        result = web_browser_integration.navigateWebPage("https://www.google.com")
        self.assertEqual(result, "https://www.google.com")

    def test_fillOutForm(self):
        form_data = {"username": "newuser", "password": "newpass"}
        result = web_browser_integration.fillOutForm(form_data)
        self.assertEqual(result, form_data)

    def test_interactWithWebApp(self):
        interaction_data = {"click": "submit_button"}
        result = web_browser_integration.interactWithWebApp(interaction_data)
        self.assertEqual(result, interaction_data)

    def test_retrieveWebPageData(self):
        result = web_browser_integration.retrieveWebPageData()
        self.assertEqual(result, self.webSessionData)

if __name__ == '__main__':
    unittest.main()
```